+++ src/frontend/apps/impress/src/features/docs/doc-editor/components/EmojiPicker.tsx
@@ -7,14 +7,12 @@ import { Box } from '@/components';
 
 interface EmojiPickerProps {
   emojiData: EmojiMartData;
-  categories: string[];
   onClickOutside: () => void;
   onEmojiSelect: ({ native }: { native: string }) => void;
 }
 
 export const EmojiPicker = ({
   emojiData,
-  categories,
   onClickOutside,
   onEmojiSelect,
 }: EmojiPickerProps) => {
@@ -24,7 +22,6 @@ export const EmojiPicker = ({
     <Box $position="absolute" $zIndex={1000} $margin="2rem 0 0 0">
       <Picker
         data={emojiData}
-        categories={categories}
         locale={i18n.resolvedLanguage}
         navPosition="none"
         onClickOutside={onClickOutside}